Handle the clicking of tiles.
Send an AJAX POST request to /clicks when a tile is clicked. If
successful, make the tile disappear. If the request fails, make the
tile red.

If every tile has been successfully clicked replace the .tiles with a
list of the most clicked tiles by sending an AJAX GET request to
/clicks.